minimum length of encoder output for CTC training
There might be consecutive repetition of symbol in the reference,
and for this the CTC alignment must put a blank in between, so the
reverse mapping of aligned symbols produces the original reference.

I realised it recently while playing with CTC aligner from torachaudio
with the noisy yodas2 dataset.

To illustrate:

"a a b c c d e f"

- len(tokens) is 8
- but, because of duplications 'a a', 'c c'
- the minimum length of encoder output is 10
- the shortest valid CTC alignment is: "a ∅  a b c ∅  c d e f"